requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
apscheduler==3.6.3
    # via python-telegram-bot
cachetools==4.2.2
    # via python-telegram-bot
certifi==2021.10.8
    # via python-telegram-bot
gitdb==4.0.9
    # via gitpython
gitpython==3.1.24
    # via -r requirements/base.in
python-telegram-bot==13.8.1
    # via -r requirements/base.in
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pytz==2021.3
    # via
    #   apscheduler
    #   python-telegram-bot
rpi-ws281x==4.3.1
    # via -r requirements/base.in
six==1.16.0
    # via apscheduler
smmap==5.0.0
    # via gitdb
tornado==6.1
    # via python-telegram-bot
typing-extensions==4.0.0
    # via gitpython
tzdata==2021.5
    # via pytz-deprecation-shim
tzlocal==4.1
    # via apscheduler

# The following packages are considered to be unsafe in a requirements file:
# setuptools
